Umbrella Shield,Umbrella Rotation Improvements, and Turret Rework
I implemented the umbrella shield so that they reflect the bullets back at the turret that shot them. Also, there was a minor bug that used to occur: when the player flipped directions, the umbrella would visually flip with the player but not flip "code wise", so the code still thought it had a right orientation when it should have a left like the player. I updated the code to reflect this change.

Also, I reworked the turrets so they use a RayCast instead of a capsule collider to detect when something is in range. I find this to be more practical and easier to manage. I decided to update this when I was working on the umbrella shield because it made things work together easier.