Consolidated the two PlayerController's into one and created a new prefab, thePlayer.
This process involved a lot of modifications to the PlayerController script, but I made backups as text files which can be found in the BACKUPS folder in the Scripts folder. I made sure that the umbrella shield functions properly with the animations and that the torso and legs should flip and move as intended.