Severe dashing improvements, easier in-air control,and fixed level loading (on restart and on a newly loaded lvl)
Dashing now has no "cooldown," but has a clamped velocity on the x-axis. This makes it more seamless to use without getting interrupted by a sudden decrease in speed. If the player is dashing in one direction and they start moving in the opposite direction, they will stop dashing and start moving normally in the desired direction.

In-air control has been improved, so the player can move in the direction (x-axis) they want to without the fan forcing them in a direction. If the player has some applied velocity in one direction on the x-axis and they start moving in the opposite direction, their x-velocity will turn to 0 and they'll be able to move freely, without any hinderance from the fan.

As for level loading, I made a new script, called the SceneLoadingController, that will run the PrepareLevel() function in the GameControllerScript, making sure that the player's lives are at max and that all necessary GameObject's are set active and updated. (noteably the LivesText).