Fixed BG music looping bug
On occasion, the intro of the main background music would play, but then the looping section would get delayed. This has been fixed.
Also, if you were to enter a level and pause the game before the loop started playing, the loop would play delayed based on how long you paused the game for. This has been fixed.
Finally, if you were to enter a level and quickly go back to the main menu before the loop began playing, the loop would play alongside the main menu music. This has been fixed!